Blocks: Move UpgradeNudge into components/
